Added Folders and moved api call to seperate method
Added folders to the page so you can view by folder. The styling for
the folders links on the page needs work. Formatting method for Folders
removes the Uncatagorized folder.

Since we are now making two API calls I moved the Guzzle API call to it's
own function to simplify things.